requirements.txt
fastapi>=0.115.0
starlette>=0.37.2
pydantic[email]>=2.8.0,<3.0.0
pydantic-core>=2.20.0
sqlmodel>=0.0.24
SQLAlchemy>=2.0.30,<3.0
psycopg[binary]==3.2.3
uvicorn[standard]>=0.30.0
greenlet>=3.2.4,<4


#pub sub dependancies
grpcio>=1.62.0
grpcio-tools>=1.62.0
avro>=1.11.3
requests>=2.32.3


#pswd encryption
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9 

